dashboard users

+++ api/app/Http/Controllers/Api/OrderMarkController.php
@@ -6,12 +6,15 @@ use App\Models\Order;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\OrderResource;
+use Illuminate\Support\Facades\Gate;
 
 class OrderMarkController extends Controller
 {
 
     public function store(Order $order)
     {
+        Gate::authorize('update', $order);
+
         if ($order->mark) {
             $order->mark()->delete();
         } else {
